+++ vehicle/OVMS.V3/components/vehicle_smarted/src/vehicle_smarted.h
@@ -66,7 +66,7 @@ class OvmsVehicleSmartED : public OvmsVehicle
     void IncomingFrameCan1(CAN_frame_t* p_frame) override;
     void IncomingFrameCan2(CAN_frame_t* p_frame) override;
     void IncomingPollReply(const OvmsPoller::poll_job_t &job, uint8_t* data, uint8_t length) override;
-    void IncomingPollError(const OvmsPoller::poll_job_t &job, uint16_t code) override;
+    void IncomingPollError(const OvmsPoller::poll_job_t &job, int32_t code) override;
     char m_vin[18];
 
   public:
@@ -357,7 +357,7 @@ class OvmsVehicleSmartED : public OvmsVehicle
   
   protected:
     string              smarted_obd_rxbuf;
-    uint16_t            smarted_obd_rxerr;
+    int32_t             smarted_obd_rxerr;
     OvmsMutex           smarted_obd_request;
     OvmsSemaphore       smarted_obd_rxwait;
